fix(ritk-snap): Preserve native coordinate precision
Use f64 at the format-neutral presentation boundary so native i32 and browser display coordinates are translated without silent rounding.

Item: RITK-SNAP-METIS-001

Refs: backlog.md#RITK-SNAP-METIS-001

+++ crates/ritk-snap/src/app/viewport_render.rs
@@ -62,7 +62,7 @@ fn client_point(point: egui::Pos2) -> Option<ViewportPoint> {
     if !point.x.is_finite() || !point.y.is_finite() {
         return None;
     }
-    Some(ViewportPoint::new(point.x, point.y))
+    Some(ViewportPoint::new(f64::from(point.x), f64::from(point.y)))
 }
 
 impl SnapApp {
